Reduce arcs for lexical rules
For potential arcs where the first symbol is lexical,
we assume that the whole rule must be lexical. Therefore,
instead of parsing multiple lexical rules simultaneously,
we just look ahead to the end of the phrase, and only yield
productions that entirely match a subset of those tokens.

This makes a particular difference in Japanese, where you
have no word boundaries and virtually all lexical productions
have several characters, but a starting character can yield
dozens of productions invalided with the second character.